Component: se agrego el componente de donutchart
